Factored out scale factor implementation and better implemented pytest.
